requirements.txt
djangorestframework
asgiref==3.2.3
Django==3.0.1
pytz==2019.3
sqlparse==0.3.0
